requirements.txt
openai-whisper
numpy
torch
python-dotenv